Use less extreme language in fun questions
Update the "fun question" language to be a bit less extreme. Asking a candidate for "*the coolest*" or "*most* proud of" may make them feel like they have to find the absolute coolest, or like they are being judged for how cool their "thing" is. The more relaxed language — "*a* cool thing", "*something* you're proud of" — can be less intimidating to candidates.